Runtime: MySQL fixes for replication cluster

+++ python/cloudtik/runtime/common/scripts/util-functions.sh
@@ -120,12 +120,14 @@ function stop_process_by_name() {
 
 function stop_process_by_pid_file() {
     local PROCESS_PID_FILE=$1
-    local PROCESS_NAME=$(basename "$PROCESS_PID_FILE")
-    local MY_PID=$(sudo pgrep --pidfile ${PROCESS_PID_FILE})
-    if [ -n "${MY_PID}" ]; then
-        echo "Stopping ${PROCESS_NAME}..."
-        # SIGTERM = 15
-        sudo kill -15 ${MY_PID} >/dev/null 2>&1
+    if [ sudo test -f "$PROCESS_PID_FILE" ]; then
+        local PROCESS_NAME=$(basename "$PROCESS_PID_FILE")
+        local MY_PID=$(sudo pgrep --pidfile ${PROCESS_PID_FILE})
+        if [ -n "${MY_PID}" ]; then
+            echo "Stopping ${PROCESS_NAME}..."
+            # SIGTERM = 15
+            sudo kill -15 ${MY_PID} >/dev/null 2>&1
+        fi
     fi
 }
